moves om testing into own test file
